+++ test/test_ecc_reedsolomon.jl
@@ -4,7 +4,7 @@ using QuantumClifford
 using QuantumClifford.ECC
 using QuantumClifford.ECC: AbstractECC, ReedSolomon
 
-@testset "Test ReedSolomon(n, k)  generator polynomial g(x) degree" begin
+@testset "Test ReedSolomon(n, k) generator polynomial g(x) degree" begin
     # RS(7, 3), RS(15, 9), RS(255, 223), RS(160, 128), RS(255, 251), (255, 239) and (255, 249) codes
     test_cases = [(7, 3), (15, 9), (225, 223), (160, 128), (255, 251), (255, 239), (255, 249)]
     for (n, k) in test_cases